Reorder service function arguments
The order of the function arguments determines the order in which errors
get returned to the client. The order was changed to `User, Path, Query,
Json, Data`. This means that Actix Web will first extract the `User`
from the query, returning 401 if it can't, followed by extracting the
path, returning 404 if it can't, etc.

Signed-off-by: Klim Tsoutsman <klimusha@gmail.com>